+++ src/index.js
@@ -1,0 +1,4 @@
+import Tippy from './Tippy'
+
+export * from './LazyTippy'
+export default Tippy
