+++ test/Tippy.test.js
@@ -1,13 +1,7 @@
 import React from 'react'
-import Tippy from '../src/Tippy'
+import Tippy from '../src'
 import ReactDOMServer from 'react-dom/server'
-import {
-  render,
-  fireEvent,
-  cleanup,
-  waitForElement,
-  wait
-} from 'react-testing-library'
+import { render, fireEvent, cleanup } from 'react-testing-library'
 
 afterEach(cleanup)
 
@@ -115,72 +109,6 @@ describe('<Tippy />', () => {
         </Tippy>
       ).includes('<div>Tooltip</div>')
     ).toBe(false)
-  })
-
-  test('tooltip content is not rendered before show when lazyRender', () => {
-    class Content extends React.Component {
-      render() {
-        throw new Error('don`t need render')
-        return ''
-      }
-    }
-
-    render(
-      <Tippy content={<Content />} lazyRender>
-        <button />
-      </Tippy>
-    )
-  })
-
-  test('tooltip content is rendered after trigger when lazyRender', async done => {
-    const Content = () => <span data-testid="content">Tooltip</span>
-
-    const onShow = jest.fn()
-    const onHidden = jest.fn()
-
-    const { container, getByTestId } = render(
-      <Tippy
-        duration={[0, 0]}
-        trigger="click"
-        onHidden={onHidden}
-        onShow={onShow}
-        content={<Content />}
-        lazyRender
-      >
-        <button />
-      </Tippy>
-    )
-
-    const button = container.querySelector('button')
-    const tip = button._tippy
-    fireEvent.click(button)
-    await waitForElement(() => getByTestId('content'))
-    fireEvent.click(button)
-
-    expect(onShow).toBeCalledTimes(1)
-    expect(onShow).toBeCalledWith(tip)
-    await wait(() => [
-      expect(onHidden).toBeCalledTimes(1),
-      expect(onHidden).toBeCalledWith(tip)
-    ])
-    done()
-  })
-
-  test('tooltip content is rendered after trigger when lazyRender without callbacks', async done => {
-    const Content = () => <span data-testid="content">Tooltip</span>
-
-    const { container, getByTestId } = render(
-      <Tippy duration={[0, 0]} trigger="click" content={<Content />} lazyRender>
-        <button />
-      </Tippy>
-    )
-
-    const button = container.querySelector('button')
-    const tip = button._tippy
-    fireEvent.click(button)
-    await waitForElement(() => getByTestId('content'))
-    fireEvent.click(button)
-    done()
   })
 
   test('props.isEnabled initially `true`', done => {
